+++ src/vpso/typing.py
@@ -1,9 +1,9 @@
-from numpy import float64, int64
+from numpy import float64, int32
 from numpy.typing import NDArray
 from typing_extensions import TypeAlias
 
 Array1d: TypeAlias = NDArray[float64]
 Array2d: TypeAlias = NDArray[float64]
 Array3d: TypeAlias = NDArray[float64]
-Array1i: TypeAlias = NDArray[int64]
-Array2i: TypeAlias = NDArray[int64]
+Array1i: TypeAlias = NDArray[int32]
+Array2i: TypeAlias = NDArray[int32]
