+++ src/vpso/vpso.py
@@ -122,7 +122,7 @@ def vpso(
     sx, sf = get_best(px, pf, nvec)  # social/global best position/value
 
     # main optimization loop
-    patience_level = np.zeros((nvec, 2), dtype=np.int64)  # 2 level for x and for f
+    patience_level = np.zeros((nvec, 2), dtype=np.int32)  # 2 level for x and for f
     termination_reason = "maxiter"
     for _ in range(maxiter):
         x, v = generate_offsprings(
